Functions

Missing docstring.

Missing docstring for create_lambda_function( remote_image::RemoteImage, role::AWSRole; function_name::Union{Nothing, String} = nothing, timeout::Int64 = 60, memory_size::Int64 = 2000, ). Check Documenter's build log for details.

Missing docstring.

Missing docstring for create_lambda_function( repo::ECRRepo, role::AWSRole; function_name::Union{Nothing, String} = nothing, image_tag::String = "latest", timeout::Int64 = 60, memory_size::Int64 = 2000, ). Check Documenter's build log for details.

Missing docstring.

Missing docstring for create_local_image( image_suffix::String, responder::AbstractResponder; aws_config::Union{Nothing, AWSConfig} = nothing, image_tag::String = "latest", no_cache::Bool = false, julia_base_version::String = "1.6.1", julia_cpu_target::String = "x86-64", package_compile::Bool = false, ). Check Documenter's build log for details.

Base.delete! — Method

delete!(con::Container)

Deletes the passed container from the local docker system. The Container instance continues to exist, but has its exists attribute set to false.

Base.delete! — Method

delete!(repo::ECRRepo)

Removes the passed ECRRepo instance from AWS ECR, and sets the exists attribute to false to indicate it no longer exists.

Base.delete! — Method

delete!(func::LambdaFunction)

Deletes a Lambda function hosted on AWS. The LambdaFunction instance continues to exist, but has its exists attribute set to false.

Base.delete! — Method

delete!(
  image::LocalImage; 
  force::Bool=false,
)

Deletes a locally-stored docker image. The LocalImage instance continues to exist, but has its exists attribute set to false.

Jot.get_dockerfile — Method

get_dockerfile(
    responder::AbstractResponder,
    julia_base_version::String,
    package_compile::Bool,
  )::String

Returns contents for a Dockerfile. This function is called in create_local_image in order to create a local docker image.

Jot.get_ecr_repo — Method

get_ecr_repo(image::LocalImage)::Union{Nothing, ECRRepo}

Queries AWS and returns an ECRRepo instance that is associated with the passed local_image. Returns nothing if one cannot be found.

Jot.get_ecr_repo — Method

get_ecr_repo(repo_name::String)::Union{Nothing, ECRRepo}

Queries AWS and returns an ECRRepo instance with the passed repo_name. Returns nothing if one cannot be found.

Jot.create_lambda_components — Method

create_lambda_components(
    res::AbstractResponder;
    image_suffix::Union{Nothing, String} = nothing,
    aws_config::Union{Nothing, AWSConfig} = nothing, 
    image_tag::String = "latest",
    no_cache::Bool = false,
    julia_base_version::String = "1.6.1",
    julia_cpu_target::String = "x86-64",
    package_compile::Bool = false,
    user_defined_labels::AbstractDict{String, String} = OrderedDict{String, String}(),
  )::LambdaComponents

Creates a LocalImage from the given responder, creates a LambdaComponents object to store the local image, and then returns the LambdaComponents object.

Acts as an alternative to create_local_image, but returns a LambdaComponents rather than just the local image. This can be more convenient for keeping the components of a lambda function organized - for example: create_lambda_components(responder) |> with_remote_image! |> with_lambda_function! will run through the entire process of creating a local image, pushing that image to ECR, and then creating a Lambda function.

Jot.get_all_aws_roles — Method

function get_all_aws_roles()::Vector{AWSRole}

Get all AWS roles from the current AWS Config.

Jot.get_all_containers — Function

get_all_containers(args::Vector{String} = Vector{String}())::Vector{Container}

Returns a list of containers currently available on the local machine.

args are additional arguments passed to the docker ps call that this function wraps.

Jot.get_all_ecr_repos — Function

get_all_ecr_repos(jot_generated_only::Bool = true)::Vector{ECRRepo}

Returns a vector of ECRRepos, representing all AWS-hosted ECR Repositories.

jot_generated_only specifies whether to filter for jot-generated repos only.

Jot.get_all_lambda_functions — Function

get_all_lambda_functions(jot_generated_only::Bool = true)::Vector{LambdaFunction}

Returns a vector of LambdaFunctions, representing all AWS-hosted Lambda functions.

jot_generated_only specifies whether to filter for jot-generated lambda functions only.

Jot.get_all_local_images — Method

get_all_local_images(;
    args::Vector{String} = Vector{String}(),
    jot_generated_only::Bool = true,
  )::Vector{LocalImage}

Returns a vector of LocalImages, representing all locally-stored docker images.

args are passed to the call to docker image ls, that is used to populate this vector. jot_generated_only specifies whether to filter for jot-generated images only.

Jot.get_all_remote_images — Function

function get_all_remote_images(jot_generated_only::Bool = true)::Vector{RemoteImage}

Returns all remote images stored on AWS ECR. By default, filters for jot-generated images only.

Jot.get_lambda_function — Method

get_lambda_function(function_name::String)::Union{Nothing, LambdaFunction}

Queries AWS and returns a LambdaFunction object, representing a Lambda Function hosted on AWS.

Jot.get_lambda_function — Method

get_lambda_function(repo::ECRRepo)::Union{Nothing, LambdaFunction}

Queries AWS and returns a LambdaFunction object, representing a Lambda Function hosted on AWS. The Lambda function returned is based off the given ECRRepo instance.

Jot.get_local_image — Method

get_local_image(
  identity::String,
)::Union{Nothing, LocalImage}

Returns a LocalImage object, representing a locally-stored docker image.

The passed identity string may be the repository name, or the docker image ID. If the image ID, it must be at least four characters in length.

Jot.get_remote_image — Method

get_remote_image(lambda_function::LambdaFunction)::RemoteImage

Queries AWS and returns a RemoteImage object, representing a docker image hosted on AWS ECR. The RemoteImage returned provides the code for the provided lambda_function.

Jot.get_remote_image — Method

get_remote_image(local_image::LocalImage)::Union{Nothing, RemoteImage}

Queries AWS and returns a RemoteImage instance corresponding to the given local_image.

If multiple valid images exist, this will return the first only. If none exists, returns nothing.

Jot.get_remote_image — Method

get_remote_image(identity::AbstractString)::Union{Nothing, RemoteImage}

Queries AWS and returns a RemoteImage instance corresponding to the given identity string.

The identity string will attempt to match on the name of the remote image, or the image's Digest.

If multiple valid images exist, this will return the first only. If none exists, returns nothing.

Jot.get_responder — Method

function get_responder(
    path_url::String, 
    response_function::Symbol,
    response_function_param_type::Type{IT};
    dependencies = Vector{String}(),
  )::AbstractResponder{IT} where {IT}

Returns an AbstractResponder, a type that holds the function that will be used to respond to AWS Lambda calls.

path_url may be either a local filesystem path, or a url.

If a filesystem path, it may point to either a script or a package. If a script, dependencies may be passed to specify any dependencies used in the script. If a package, the dependencies will be found automatically from its Project.toml.

If a url, it should be a remote package, for example the URL for a github repo. The url given will be passed to Pkg as a url, so any url valid in a PackageSpec will also be valid here, such as https://github.com/harris-chris/JotTest3

response_function is a function within this module that you would like to use to respond to AWS Lambda calls. response_function_param_type specifies the type that the response function is expecting as its only argument.

Jot.get_responder — Method

function get_responder(
    package_spec::Pkg.Types.PackageSpec, 
    response_function::Symbol,
    response_function_param_type::Type{IT};
    dependencies = Vector{String}(),
  )::AbstractResponder{IT} where {IT}

Returns an AbstractResponder, a type that holds the function that will be used to respond to AWS Lambda calls.

package_spec is an instance of PackageSpec, part of the standard Julia Pkg library.

response_function is a function within this module that you would like to use to respond to AWS Lambda calls. response_function_param_type specifies the type that the response function is expecting as its only argument.

Jot.get_responder — Method

function get_responder(
    mod::Module, 
    response_function::Symbol,
    response_function_param_type::Type{IT},
  )::AbstractResponder{IT} where {IT}

Returns an AbstractResponder, a type that holds the function that will be used to respond to AWS Lambda calls.

mod is a module currently in scope, and response_function is a function within this module that you would like to use to respond to AWS Lambda calls.

response_function is a function within this module that you would like to use to respond to AWS Lambda calls. response_function_param_type specifies the type that the response function is expecting as its only argument.

Missing docstring.

Missing docstring for get_user_labels(l::Union{LambdaComponent, ECRRepo}). Check Documenter's build log for details.

Jot.invoke_function — Method

invoke_function(
    request::Any,
    lambda_function::LambdaFunction;
    check_state::Bool=false,
  )::Any

Invokes a Lambda function, hosted on AWS. request is the argument that it will be called with. This will be automatically converted to JSON before sending, so should match the response_function_param_type of the responder used to create the function.

Returns the invoked Lambda function response, or throws an error if the invoked Lambda function has returned an error status.

If check_state is true, the function will wait for the AWS Lambda function to become available before sending the request. This can be useful if the Lambda function has been created within the last few seconds, since there is a short set-up time before it can be called.

Jot.is_container_running — Method

is_container_running(con::Container)::Bool

Returns true if the given docker container is currently running (not stopped).

Jot.push_to_ecr! — Method

push_to_ecr!(image::LocalImage)::RemoteImage

Pushes the given local docker image to an AWS ECR Repo, a prerequisite of creating an AWS Lambda Function. If an ECR Repo for the given local image does not exist, it will be created automatically. Returns a RemoteImage object that represents the docker image that is hosted on the ECR Repo. The ECR Repo itself is an attribute of the RemoteImage.

Jot.run_image_locally — Method

run_image_locally(local_image::LocalImage; detached::Bool=true)::Container

Runs the given local image, starting a docker container. If detached, the container will run in the background. The container can be stopped/deleted by eg stop_container, delete!.

Jot.run_test — Function

run_test(
  image::LocalImage,
  function_argument::Any = "", 
  expected_response::Any = nothing;
  then_stop::Bool = false,
)::Tuple{Bool, Float64}

Runs a test of the given local docker image, passing function_argument (if given), and expecting expected_response(if given). If a function_argument is not given, then it will merely test that any kind of response is received - this response may be an error JSON and the test will still pass, establishing only that the image can be contacted. Returns a tuple of (test result, time) where time is the time taken for a response to be received, in seconds.

The test will use an already-running docker container, if one exists. If this is the case then the then_stop parameter tells the function whether to stop the docker container after running the test. If the runtest function finds no docker container already running, it will start one, and then shut it down afterwards, regardless of the value of `thenstop`.

Jot.run_test — Method

run_test(
    l::LambdaComponents;
    function_argument::Any = "", 
    expected_response::Any = nothing;
  )::Tuple{Bool, Float64}

Tests the passed LambdaComponents instance.

The test runs on the most downstream object. So if the instance has a LambdaFunction, this will be tested. Otherwise, the attached LocalImage will be tested. If the LambdaComponents object has neither a local image or a lambda function, then it has nothing that can be tested and the function will throw an error.

Returns a tuple of {Test pass/fail, Test time taken in seconds}.

Jot.send_local_request — Method

send_local_request(request::Any)

Make a function call to a locally-running docker container and returns the value. A container can be initiated by eg run_image_locally.

Jot.show_lambdas — Method

show_lambdas()::Nothing

Displays a table of all objects generated using Jot.jl.

Each row of the table shows at least one of a Responder, a local docker image, a remote (hosted on AWS ECR) docker image, and an AWS-hosted Lambda function. The local docker image, remote docker image and lambda function on a given row of the table are guaranteed to share the same underlying function code.

The Responder column is colour-coded:

Grey indicates that this path no longer exists.

White indicates that this path still exists, but the code has changed since the objects shown

in the row were created.

Blue indicates that the underlying code for this row (eg the code present in the local image,

remote image etc) is the same as is currently present at this path.

Jot.stop_container — Method

stop_container(con::Container)

Stops the given docker container, if currently running.

Jot.with_remote_image! — Method

with_remote_image!(l::LambdaComponents)::LambdaComponents

Adds a 'RemoteImageobject to the passedLambdaComponents` instance. Will error if the instance has neither an existing remote image or local image.

Jot.with_lambda_function! — Method

with_lambda_function!(l::LambdaComponents)::LambdaComponents

Adds a 'LambdaFunctioninstance to the passedLambdaComponents` instance. Will error if the instance has neither an existing remote image, local image or lambda function.