Functions

Jot.create_lambda_function — Method

create_lambda_function(
    remote_image::RemoteImage,
    role::AWSRole;
    function_name::Union{Nothing, String} = nothing,
    timeout::Int64 = 60,
    memory_size::Int64 = 2000,
  )::LambdaFunction

Creates a function that exists on the AWS Lambda service. The function will use the given RemoteImage, and runs using the given AWS Role.

Jot.create_lambda_function — Method

create_lambda_function(
    repo::ECRRepo,
    role::AWSRole;
    function_name::Union{Nothing, String} = nothing,
    image_tag::String = "latest",
    timeout::Int64 = 60,
    memory_size::Int64 = 2000,
  )::LambdaFunction

Creates a function that exists on the AWS Lambda service. The function will use the given ECR Repo, and runs using the given AWS Role. If given, the image_tag will decide which of the images in the ECR Repo is used.

Jot.create_local_image — Method

create_local_image(
    image_suffix::String,
    responder::AbstractResponder;
    aws_config::Union{Nothing, AWSConfig} = nothing, 
    image_tag::String = "latest",
    no_cache::Bool = false,
    julia_base_version::String = "1.6.1",
    julia_cpu_target::String = "x86-64",
    package_compile::Bool = false,
  )::LocalImage

Creates a locally-stored docker image containing the specified responder. This can be tested tested locally, or directly uploaded to an AWS ECR Repo for use as an AWS Lambda function.

package_compile indicates whether PackageCompiler.jl should be used to compile the image - this is not necessarily for testing/exploration but is highly recommended for production use.

Use no_cache to construct the local image without using a cache; this is sometimes necessary if nothing locally has changed, but the image is querying a remote object (say, a github repo) which has changed.

Base.delete! — Method

delete!(con::Container)

Deletes the passed container from the local docker system. The Container instance continues to exist, but has its exists attribute set to false.

Base.delete! — Method

delete!(repo::ECRRepo)

Removes the passed ECRRepo instance from AWS ECR, and sets the exists attribute to false to indicate it no longer exists.

Base.delete! — Method

delete!(func::LambdaFunction)

Deletes a Lambda function hosted on AWS. The LambdaFunction instance continues to exist, but has its exists attribute set to false.

Base.delete! — Method

delete!(
  image::LocalImage; 
  force::Bool=false,
)

Deletes a locally-stored docker image. The LocalImage instance continues to exist, but has its exists attribute set to false.

Jot.get_dockerfile — Method

get_dockerfile(
    responder::AbstractResponder,
    julia_base_version::String,
    package_compile::Bool,
  )::String

Returns contents for a Dockerfile. This function is called in create_local_image in order to create a local docker image.

Jot.get_ecr_repo — Method

get_ecr_repo(image::LocalImage)::Union{Nothing, ECRRepo}

Queries AWS and returns an ECRRepo instance that is associated with the passed local_image. Returns nothing if one cannot be found.

Jot.get_ecr_repo — Method

get_ecr_repo(repo_name::String)::Union{Nothing, ECRRepo}

Queries AWS and returns an ECRRepo instance with the passed repo_name. Returns nothing if one cannot be found.

Jot.get_lambda_function — Method

get_lambda_function(function_name::String)::Union{Nothing, LambdaFunction}

Queries AWS and returns a LambdaFunction object, representing a Lambda Function hosted on AWS.

Jot.get_lambda_function — Method

get_lambda_function(repo::ECRRepo)::Union{Nothing, LambdaFunction}

Queries AWS and returns a LambdaFunction object, representing a Lambda Function hosted on AWS. The Lambda function returned is based off the given ECRRepo instance.

Jot.get_local_image — Method

get_local_image(
  repository::String,
)::Union{Nothing, LocalImage}

Returns a LocalImage object, representing a locally-stored docker image.

Jot.get_remote_image — Method

get_remote_image(lambda_function::LambdaFunction)::RemoteImage

Queries AWS and returns a RemoteImage object, representing a docker image hosted on AWS ECR. The RemoteImage returned provides the code for the provided lambda_function.

Jot.get_remote_image — Method

get_remote_image(local_image::LocalImage)::Union{Nothing, RemoteImage}

Queries AWS and returns a RemoteImage instance corresponding to the given local_image. If none exists, returns nothing.

Jot.get_responder — Method

function get_responder(
    path_url::String, 
    response_function::Symbol,
    response_function_param_type::Type{IT};
    dependencies = Vector{String}(),
  )::AbstractResponder{IT} where {IT}

Returns an AbstractResponder, a type that holds the function that will be used to respond to AWS Lambda calls.

path_url may point to either a script or a package. If a script, dependencies may be passed to specify any dependencies used in the script. If a package, the dependencies will be found automatically from its Project.toml.

response_function is a function within this module that you would like to use to respond to AWS Lambda calls. response_function_param_type specifies the type that the response function is expecting as its only argument.

Jot.get_responder — Method

function get_responder(
    package_spec::Pkg.Types.PackageSpec, 
    response_function::Symbol,
    response_function_param_type::Type{IT};
    dependencies = Vector{String}(),
  )::AbstractResponder{IT} where {IT}

Returns an AbstractResponder, a type that holds the function that will be used to respond to AWS Lambda calls.

package_spec is an instance of PackageSpec, part of the standard Julia Pkg library.

response_function is a function within this module that you would like to use to respond to AWS Lambda calls. response_function_param_type specifies the type that the response function is expecting as its only argument.

Jot.get_responder — Method

function get_responder(
    mod::Module, 
    response_function::Symbol,
    response_function_param_type::Type{IT},
  )::AbstractResponder{IT} where {IT}

Returns an AbstractResponder, a type that holds the function that will be used to respond to AWS Lambda calls.

mod is a module currently in scope, and response_function is a function within this module that you would like to use to respond to AWS Lambda calls.

response_function is a function within this module that you would like to use to respond to AWS Lambda calls. response_function_param_type specifies the type that the response function is expecting as its only argument.

Jot.invoke_function — Method

invoke_function(
    request::Any,
    lambda_function::LambdaFunction;
    check_state::Bool=false,
  )::Any

Invokes a Lambda function, hosted on AWS. request is the argument that it will be called with. This will be automatically converted to JSON before sending, so should match the response_function_param_type of the responder used to create the function.

Returns the invoked Lambda function response, or throws an error if the invoked Lambda function has returned an error status.

If check_state is true, the function will wait for the AWS Lambda function to become available before sending the request. This can be useful if the Lambda function has been created within the last few seconds, since there is a short set-up time before it can be called.

Jot.is_container_running — Method

is_container_running(con::Container)::Bool

Returns true if the given docker container is currently running (not stopped).

Jot.push_to_ecr! — Method

push_to_ecr!(image::LocalImage)::Tuple{ECRRepo, RemoteImage}

Pushes the given local docker image to an AWS ECR Repo, a prerequisite of creating an AWS Lambda Function. If an ECR Repo for the given local image does not exist, it will be created automatically. Returns both the ECR Repo, and a RemoteImage object that represents the docker image that is hosted on the ECR Repo.

Jot.run_image_locally — Method

run_image_locally(local_image::LocalImage; detached::Bool=true)::Container

Runs the given local image, starting a docker container. If detached, the container will run in the background. The container can be stopped/deleted by eg stop_container, delete!.

Jot.run_test — Function

run_test(
  image::LocalImage,
  function_argument::Any = "", 
  expected_response::Any = nothing;
  then_stop::Bool = false,
)::Tuple{Bool, Float64}

Runs a test of the given local docker image, passing function_argument (if given), and expecting expected_response(if given). If a function_argument is not given, then it will merely test that any kind of response is received - this response may be an error JSON and the test will still pass, establishing only that the image can be contacted. Returns a tuple of (test result, time) where time is the time taken for a response to be received, in seconds.

The test will use an already-running docker container, if one exists. If this is the case then the then_stop parameter tells the function whether to stop the docker container after running the test. If the runtest function finds no docker container already running, it will start one, and then shut it down afterwards, regardless of the value of `thenstop`.

Jot.send_local_request — Method

send_local_request(request::Any)

Make a function call to a locally-running docker container and returns the value. A container can be initiated by eg run_image_locally.

Jot.stop_container — Method

stop_container(con::Container)

Stops the given docker container, if currently running.