Types

Jot.AbstractResponder — Type

abstract type AbstractResponder{IT} end

The supertype of all responder types. The type parameter represents the parameter type of the response_function responder attribute.

Jot.AWSConfig — Type

mutable struct AWSConfig
    account_id::Union{Missing, String} = missing
    region::Union{Missing, String} = missing
end

Defines the Amazon Web Services account id and region to use.

Your current AWS CLI profile should match the account id specified here. To check that this is the case, run aws sts get-caller-identity --query Account --output text from the command line. To see available profiles, run aws configure list-profiles.

Jot.Container — Type

mutable struct Container
    ID::String
    Image::String
    Command::Union{Missing, String} = missing
    CreatedAt::Union{Missing, String} = missing
    Names::Union{Missing, String} = missing
    Ports::Union{Missing, String} = missing
    exists::Bool = true
end

Represents a docker container on the local environment. Should not be instantiated directly. If exists is true, then the container is assumed to exit and so should be visible from utilities such as docker container ls --all.

Jot.LambdaFunction — Type

@with_kw mutable struct LambdaFunction
    FunctionName::Union{Missing, String} = missing
    FunctionArn::Union{Missing, String} = missing
    Runtime::Union{Missing, String} = missing
    Role::Union{Missing, String} = missing
    Handler::Union{Missing, String} = missing
    CodeSize::Union{Missing, Int64} = missing
    Description::Union{Missing, String} = missing
    Timeout::Union{Missing, Int64} = missing
    MemorySize::Union{Missing, Int64} = missing
    LastModified::Union{Missing, String} = missing
    CodeSha256::Union{Missing, String} = missing
    Version::Union{Missing, String} = missing
    TracingConfig::Union{Missing, Dict{String, Any}} = missing
    RevisionId::Union{Missing, String} = missing
    PackageType::Union{Missing, String} = missing
    exists::Bool = true
end

Represents a Lambda function, hosted on AWS. Should not be instantiated directly. If exists is true, then the image is assumed to exit and so should be visible from utilities such as aws lambda list-functions

Jot.LocalImage — Type

mutable struct LocalImage
    CreatedAt::Union{Missing, String} = missing
    Digest::String
    ID::String
    Repository::String
    Size::Union{Missing, String} = missing
    Tag::String
    exists::Bool = true
end

Represents a docker image on the local machine, and stores associated metadata. Should not be instantiated directly. If exists is true, then the image is assumed to exit and so should be visible from utilities such as docker image ls.

Jot.LocalPackageResponder — Type

struct LocalPackageResponder{IT} <: AbstractResponder{IT}
    pkg::Pkg.Types.PackageSpec
    original_path::String
    response_function::Symbol
    response_function_param_type::Type{IT}
    build_dir::String
    package_name::String
end

A responder that is located locally (in the temporary build_dir) and is a Julia package. This is usually created by the Responder function.

Jot.RemoteImage — Type

@with_kw mutable struct RemoteImage
    imageDigest::Union{Missing, String} = missing
    imageTag::Union{Missing, String} = missing
    ecr_repo::Union{Missing, ECRRepo} = missing
    exists::Bool = true
end

Represents a docker image stored in an AWS ECR Repository. The exists attribute indicates whether the RemoteImage still exists.